environments/coffee/pyproject.toml
[project]
name = "coffee-marketplace"
version = "0.1.0"
description = "Multi-agent coffee bean marketplace simulation"
requires-python = ">=3.9"
dependencies = [
    "PyYAML",
    "google-genai",
    "python-dotenv",
    "openai",
    "azure-identity",
    "numpy",
    "pandas",
    "matplotlib",
]

[project.optional-dependencies]
vllm = [
    "vllm>=0.11.0",
    "torch>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
